+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 CHANGELOG for Sulu
 ==================
+* dev-master
+    * FEATURE     #3028 [MediaBundle]             Added blur, grayscale, gamma, negative and sharpen transformation to media
+    * HOTFIX      #3750 [PreviewBundle]           Fixed refresh preview
 
     * HOTFIX      #3747 [RouteBundle]             Added empty array as default value to histories property.